🖨 GCode improvements for contour printing

+++ Slicer/slyce/GCode/Commands/M/FanOn.cs
@@ -6,7 +6,7 @@ namespace Slicer.slyce.GCode.Commands
         [ParameterType("P")]
         public int? FanNumber { get; set; }
         [ParameterType("S")]
-        public double FanSpeed { get; set; }
+        public int FanSpeed { get; set; }
         [ParameterType("I")]
         public int? Invert { get; set; }
         [ParameterType("F")]
